requirements/base.txt
fastapi[standard]~=0.115.14
fastapi-utils[all]~=0.8.0

sqlmodel~=0.0.24
psycopg2~=2.9.10
asyncpg~=0.30.0

pydantic~=2.11.7
pydantic-settings~=2.10.1

pyjwt[crypto]~=2.10.1
passlib[bcrypt]~=1.7.4
